%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.NullFileLock}
\alias{getConnection}
\alias{getConnection.default}
\alias{getConnection.FileLock}
\alias{getPathname}
\alias{getPathname.FileLock}
\alias{isLocked}
\alias{isLocked.FileLock}
\alias{lock.FileLock}
\alias{NullFileLock}
\alias{release}
\alias{release.FileLock}
\alias{tryLock.FileLock}
\alias{as.character.FileLock}
\alias{finalize.FileLock}
\alias{lock}
\alias{tryLock}
\alias{acquire.AbstractMutex}
\alias{finalize.AbstractMutex}
\alias{finalize.SocketMutex}
\alias{getHost}
\alias{getHost.SocketMutex}
\alias{getPort}
\alias{getPort.SocketMutex}
\alias{getPort.SocketMutexServer}
\alias{isAcquired.AbstractMutex}
\alias{isAcquired.NullMutex}
\alias{isAcquired.SocketMutex}
\alias{release.AbstractMutex}
\alias{release.NullMutex}
\alias{release.SocketMutex}
\alias{run.SocketMutexServer}
\alias{SocketMutexServer}
\alias{tryAcquire.AbstractMutex}
\alias{tryAcquire.NullMutex}
\alias{tryAcquire.SocketMutex}
\alias{acquire}
\alias{isAcquired}
\alias{run}
\alias{tryAcquire}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
